Extract TypeConstructionHelper
