+++ test/type_construction_test.rb
@@ -5,6 +5,7 @@ class TypeConstructionTest < Minitest::Test
   include TypeErrorAssertions
   include FactoryHelper
   include SubtypingHelper
+  include TypeConstructionHelper
 
   Namespace = RBS::Namespace
 
@@ -59,50 +60,6 @@ end
     end
 
     super(*files, &block)
-  end
-
-  def with_standard_construction(checker, source)
-    self_type = parse_type("::Object")
-
-    annotations = source.annotations(block: source.node, factory: checker.factory, current_module: Namespace.root)
-    const_env = ConstantEnv.new(factory: factory,
-                                context: [Namespace.root])
-    type_env = TypeEnv.build(annotations: annotations,
-                             subtyping: checker,
-                             const_env: const_env,
-                             signatures: checker.factory.env)
-    lvar_env = LocalVariableTypeEnv.empty(
-      subtyping: checker,
-      self_type: self_type
-    ).annotate(annotations)
-
-    context = Context.new(
-      block_context: nil,
-      method_context: nil,
-      module_context: Context::ModuleContext.new(
-        instance_type: AST::Builtin::Object.instance_type,
-        module_type: AST::Builtin::Object.module_type,
-        implement_name: nil,
-        current_namespace: Namespace.root,
-        const_env: const_env,
-        class_name: AST::Builtin::Object.module_name,
-        instance_definition: checker.factory.definition_builder.build_instance(AST::Builtin::Object.module_name),
-        module_definition: checker.factory.definition_builder.build_singleton(AST::Builtin::Object.module_name)
-      ),
-      break_context: nil,
-      self_type: self_type,
-      type_env: type_env,
-      lvar_env: lvar_env
-    )
-    typing = Typing.new(source: source, root_context: context)
-
-    construction = TypeConstruction.new(checker: checker,
-                                        source: source,
-                                        annotations: annotations,
-                                        context: context,
-                                        typing: typing)
-
-    yield construction, typing
   end
 
   def test_lvar_with_annotation
@@ -2562,11 +2519,6 @@ end
         end
       end
     end
-  end
-
-  def assert_no_error(typing)
-    assert_instance_of Typing, typing
-    assert_operator typing.errors.map {|e| StringIO.new().tap {|io| e.print_to(io) }.string }, :empty?
   end
 
   def test_void2
